lib/recordio: packed writer
Test Plan:
New unit tests.

Note: The compressed, packed file produced by this writer is slightly different
from the fixture generated by the Go code. If we expect some non-determinism in
the compression algorithm, I think the unit test in this revision is sound
because it validates the writer using the reader and the reader is separately
checked against the Go implementation with its own unit tests. However, if we expect
compression to be completely deterministic, I need to debug.

Reviewers: ysaito

Reviewed By: ysaito

Subscribers: cnicolaou

Differential Revision: https://phabricator.grailbio.com/D9427

fbshipit-source-id: d29dbf5