bio/align/mapper: recordio-based index format
Summary:
Modifies the index binary format to use the new C++ recordio library. Note that
the index will not be a valid recordio file; instead, it contains two recordio
files (plus a trailer at the end) and uses the library to read them. Leveraging
the recordio library, file blocks are now compressed and reading is still
sharded and parallel.

Size reduction compared to the current index and D9260 (varint):

|format               |standard (GiB)|reduction|methylation (GiB)|reduction
|---------------------|--------------|---------|-----------------|---------
|current              |56            |0%       |74               |0%
|varint               |34            |39%      |55               |26%
|recordio, individual |31            |45%      |57               |23%
|current gzipped      |26            |54%      |52               |30%
|**recordio, grouped**|**25**        |**55%**  |**50**           |**32%**
|varint gzipped       |25            |55%      |48               |35%

~~The recordio format reduces index size by 45% / 23% (standard / methylation) vs.
55% / 35% for gzipped varint. I think the savings are pretty good, and certainly
worthwhile. The gains aren't perfect, though, because the recordio format does
not control the grouping of entries into file blocks so it stores hash table
index per entry rather than per block, which wastes some space. I compensated
slightly by not writing anything for entries that are empty, which is now
possible.~~

I've modified the hash table entries format to store groups of 1000 entries as a
single recordio item (in the packed format). This reduces the overhead of
storing hash table indices from worst case ~1/3 to less than 1/1000, while still
letting the recordio format handle block framing and compression. It's slightly
more complicated, but now the new format is even smaller than the current index
after gzip, and is only slightly larger than the gzipped varint index from
D9260.

Also, in the past few weeks, Yaz has revised the recordio format and written new
Go readers and writers and a new C++ reader. Eventually when a new C++ writer
is written, we can switch to that. The formats aren't compatible but the code is
(it detects the old format and reads it appropriately), so whenever that's ready
we can just generate a new index and switch over at the next pipeline release.

Test Plan:
Existing serialization unit tests, as well as execution tests that construct an
index and align a read.

Reviewers: sgross

Reviewed By: sgross

Subscribers: ysaito, #methylation_analysis_pipeline_development

Differential Revision: https://phabricator.grailbio.com/D9902

fbshipit-source-id: 09f3706