bazel: delete -fPIC where manually specified
Summary:
We used to specify `-fPIC` in all cc builds which were dependencies of R
packages because that is a requirement for linking in shared libraries,
and bazel did not provide a way to specify a dependency on a PIC .a
library.

With the resolution of https://github.com/grailbio/rules_r/issues/31,
R rules for bazel now explicitly state their preference for a PIC build,
and so we don't have to force all builds to be PIC.

This also allows us to use the default `cc_proto_library` rules instead
of the experimental ones from the protobuf repo.

Test Plan:
R package builds and tests pass. They used to fail before without
`-fPIC`.

Reviewers: ysaito

Reviewed By: ysaito

Maniphest Tasks: T12955

Differential Revision: https://phabricator.grailbio.com/D31313

fbshipit-source-id: caafb4a